+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - consent page filters items based on person
+- Warn about unsaved changes when logging out
+- pytest-retry since tests often failing on github (we belief due to low system/cpu performance in CI)
 
 ### Changes